requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
aiohttp-cors==0.7.0
alembic==0.9.3.dev0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.4.0
blinker==1.4
Brlapi==0.6.6
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
distro==1.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.10.1
django-crispy-forms==1.4.0
django-durationfield==0.5.1
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.0.12
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gns3-gui==2.2.17
gns3-server==2.2.17
gunicorn==20.1.0
httplib2==0.9.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==2.0.0
importlib-resources==3.0.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
ldap3==2.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
multidict==4.7.6
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
openshot-qt==2.4.1
paramiko==2.0.0
passlib==1.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.6.7
psycopg2==2.9.3
py-cpuinfo==7.0.0
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
pyrsistent==0.18.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.4.2
python-debian==0.1.32
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.27.1
requests-unixsocket==0.1.5
scour==0.36
SecretStorage==2.3.1
sentry-sdk==1.3.1
simplejson==3.13.2
six==1.16.0
speaklater==1.4
SQLAlchemy==1.1.11
sqlparse==0.4.2
ssh-import-id==5.7
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
tldextract==1.6
typing-extensions==4.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.3
virtualenv==20.0.33
wadllib==1.3.2
Werkzeug==0.16.0
whitenoise==5.3.0
WTForms==2.1
xkit==0.0.0
yarl==1.6.3
zipp==3.3.0
zope.interface==4.3.2
